refactor: consolidate /resources into /sources, delete broken E874
Problem 1 from Discussion #3499: three overlapping public resource pages.

- Delete E874 (content/docs/tools/resources.mdx) — broken ResourcesIndex stub
- Remove ResourcesIndex from stub component list in mdx-components.tsx
- Redirect /resources → /sources (keep /resources/[id] detail pages)
- Merge resource type breakdown + enrichment stats into /sources page
- Update all links: homepage, search, sources-overview, resource detail back-link

/sources is now the single canonical entry point for browsing resources
and publications. Individual resource detail pages remain at /resources/[id].
