+++ apps/web/src/app/page.tsx
@@ -119,11 +119,11 @@ const STAT_DEFS: StatDef[] = [
     href: "/wiki",
   },
   {
-    label: "Resources",
+    label: "Sources",
     icon: ExternalLink,
     iconColor: "text-orange-500",
     getValue: (ctx) => ctx.resourceCount,
-    href: "/resources",
+    href: "/sources",
   },
 ];
 
@@ -274,11 +274,11 @@ export default function Home() {
               <span className="text-xs text-muted-foreground/40 tabular-nums">{totalPages}</span>
             </Link>
             <Link
-              href="/resources"
+              href="/sources"
               className="inline-flex items-center gap-1.5 px-3 py-1.5 rounded-md border border-border/60 text-sm text-foreground hover:bg-muted/50 transition-colors no-underline bg-background/60"
             >
               <ExternalLink size={13} className="text-orange-500" />
-              <span>Resources</span>
+              <span>Sources</span>
               <span className="text-xs text-muted-foreground/40 tabular-nums">{resourceCount}</span>
             </Link>
           </div>
