+++ apps/web/src/app/search/search-client.tsx
@@ -666,11 +666,11 @@ function BrowseState({ data }: { data: BrowseData }) {
             <span className="text-xs text-muted-foreground/40 tabular-nums">{data.totalPages}</span>
           </Link>
           <Link
-            href="/resources"
+            href="/sources"
             className="inline-flex items-center gap-1.5 px-3 py-1.5 rounded-md border border-border/60 text-sm text-foreground hover:bg-muted/50 transition-colors"
           >
             <ExternalLink size={13} className="text-orange-500" />
-            <span>Resources</span>
+            <span>Sources</span>
           </Link>
         </div>
       </section>
