+++ apps/web/src/app/sources/sources-overview-content.tsx
@@ -12,7 +12,7 @@ export function SourcesOverviewContent() {
   ).length;
 
   const stats = [
-    { label: "Resources", value: resources.length, href: "/resources" },
+    { label: "Resources", value: resources.length, href: "/sources" },
     { label: "Publications", value: publications.length, href: "/publications" },
     { label: "Peer-Reviewed Venues", value: peerReviewed },
     { label: "With Summaries", value: withSummary },
@@ -54,7 +54,7 @@ export function SourcesOverviewContent() {
 
       <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
         <Link
-          href="/resources"
+          href="/sources"
           className="group block rounded-xl border border-border/60 bg-card p-6 no-underline transition-all hover:shadow-md hover:border-border"
         >
           <h3 className="text-lg font-bold mb-2 group-hover:text-primary transition-colors">
