--- content/docs/tools/resources.mdx
@@ -1,16 +1,0 @@
----
-wikiId: E874
-title: External Resources
-description: Browse all external sources referenced in the knowledge base - papers, books, blogs, and more
-sidebar:
-  order: 3
-summary: This is a navigational index page that displays external resources (papers, books, blogs) referenced throughout the knowledge base with filtering and search functionality. It contains no substantive content itself, serving purely as infrastructure for accessing bibliographic references.
-clusters: []
-tableOfContents: false
----
-import {ResourcesIndex} from '@components/wiki';
-
-
-Browse all external resources (papers, books, blog posts, reports, etc.) referenced throughout the knowledge base. Use the filters to find resources by type or search by title/author.
-
-<ResourcesIndex showSearch={true} showFilters={true} />
